split route modules
